requirements.txt
certifi==2020.6.20
chardet==3.0.4
falcon==2.0.0
falcon-multipart==0.2.0
gunicorn==20.0.4
idna==2.10
PyJWT==1.7.1
requests==2.24.0
urllib3==1.25.10
